+++ server/api/products.js
@@ -3,12 +3,11 @@ const { Product } = require('../db/models');
 
 // GET /products
 productRouter.get('/', (req, res, next) => {
-  Product.findAll({
-
-  })
+  Product.findAll()
   .then(products => res.json(products))
   .catch(next);
 });
+
 productRouter.get('/:id', (req, res, next)=> {
   Product.findById(req.params.id)
   .then(id => res.json(id))
